rename: section is now article, with proper semantic
